Updated Video and Social Media Fields
Video: append vimeo ID from URL. Requires “video-url” field. Is there a
better way to do this? Unclear how we’re hosting videos, or if we need
this.

Social Media: Links are too long, so I added a second field
(‘social_media_site’) for the name of the social media website.